Rename Apply component to StartPage
