+++ src/containers/StartPage/StepsDescription/StepDescription.js
