+++ src/containers/StartPage/StepsDescription/StepsDescription.js
