+++ src/containers/StartPage/StepsHeader/Arrow.js
