+++ src/containers/StartPage/StepsHeader/Step.js
