+++ src/containers/StartPage/StepsHeader/StepsHeader.js
