+++ src/containers/index.js
@@ -1,8 +1,8 @@
 export { default as App } from './App/App'
-export { default as Apply } from './Apply/Apply'
 export { default as Cloud9Setup } from './Cloud9Setup/Cloud9Setup'
 export { default as HackbotNewTeam } from './HackbotNewTeam/HackbotNewTeam'
 export { default as LeaderIntake } from './LeaderIntake/LeaderIntake'
 export { default as NotFound } from './NotFound/NotFound'
 export { default as OneOffFormWrapper } from './OneOffFormWrapper/OneOffFormWrapper'
 export { default as RedeemTechDomain } from './RedeemTechDomain/RedeemTechDomain'
+export { default as StartPage } from './StartPage/StartPage'
